pyproject.toml
[project]
name = "oanda-forex-scalping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    'oandapyV20',
    "pandas>=2.3.1",
    'pandas_ta',
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
